Move single-use functions alongside where they're used
